Move Database & Connection Test Traits
