+++ tests/unit/Hooks/Database/DatabaseDriverTestCase.php
@@ -21,10 +21,10 @@ use Illuminate\Database\DatabaseManager;
 use Illuminate\Database\QueryException;
 use PDO;
 use Tenancy\Affects\Connection\Provider as ConnectionProvider;
-use Tenancy\Affects\Connection\Support\InteractsWithConnections;
+use Tenancy\Testing\Concerns\InteractsWithConnections;
 use Tenancy\Facades\Tenancy;
 use Tenancy\Hooks\Database\Provider as DatabaseProvider;
-use Tenancy\Hooks\Database\Support\InteractsWithDatabases;
+use Tenancy\Testing\Concerns\InteractsWithDatabases;
 use Tenancy\Hooks\Migrations\Provider as MigrationsProvider;
 use Tenancy\Hooks\Migrations\Support\InteractsWithMigrations;
 use Tenancy\Identification\Contracts\ResolvesTenants;
